Flush the audio sink if NuPlayerRenderer is paused during a flush
If the NuPlayerRenderer is paused when the audio is flushed, some
audio will still be in the audio sink buffer. It will be played
when the sink is resumed. If the renderer is flushed because of a
seek, the old audio will be heard before the audio from the new
position is played.

Change-Id: I016ea352903104ba1bc5f5586aa46012c52427c1